tests: Port sched_cls_drop test
Move the test under gadgets/ and use the new framework for it. The previous
test what checking the client wasn't able to connect to the client, but with
the existing testing framework it's difficult to do, so modify the testing
gadget to emit some events that are caught by the test.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>