gadgetCtx: add IsClient
Some operators are used both server-side and client-side. Soon, the
ustack operator will also be able to operator client-side and
server-side.

The IsClient method allows to detect which side they are running and
behave differently. This is similar to IsRemoteCall to detect if we are
on a server.

Command           | IsClient | IsRemoteCall |
------------------|----------|--------------|
ig run            | false    | false        |
ig image inspect  | false    | false        |
ig daemon         | false    | true         |
gadgetctl run     | true     | false (*)    |

*: Note that gadgetctl can display logs from the server, so gadgetctl
will print both lines:
    client=true, remote_call=false
    client=false, remote_call=true
Depending if the log was generated from the server or the client.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>